+++ MSA_UI.lua
@@ -1,45 +1,46 @@
--- For the UI Frames of Mass Crafting Assist adddon
+-- For the UI Frames of Mass Salvage Assist adddon
 
 local UI = {};
-MCA.UI = UI;
+MSA.UI = UI;
 
 -- Holds GUI frame load configuration details
 UI.LoadUI = function()
 
-    if not MCA.Crafting.Profs then
-        MCA.Crafting.Establish_Spells()
+    if not MSA.Crafting.Profs then
+        MSA.Crafting.Establish_Spells()
     end
 
     -- Checkbox to be placed on inscription window
-    if not UI.MMA_checkbox then
+    if not UI.MSA_checkbox then
 
-        UI.MMA_checkbox = CreateFrame ( "CheckButton" , "MMA_checkbox" , ProfessionsFrame.CraftingPage , "InterfaceOptionsCheckButtonTemplate" )
-        UI.MMA_checkbox.value = MCA_save.non_stop
-        UI.MMA_checkbox:SetChecked ( UI.MMA_checkbox.value )
-
-        UI.MMA_checkbox:SetPoint ( "TOPLEFT" , ProfessionsFrame.CraftingPage.SchematicForm , "BOTTOMLEFT" , 10 , -4 )
+        UI.MSA_checkbox = CreateFrame ( "CheckButton" , "MSA_checkbox" , ProfessionsFrame.CraftingPage , "InterfaceOptionsCheckButtonTemplate" )
+        UI.MSA_checkbox.value = MSA_save.non_stop
+        UI.MSA_checkbox:SetChecked ( UI.MSA_checkbox.value )
 
         -- Text to the right of checkbox
-        UI.MMA_checkbox.Text = UI.MMA_checkbox:CreateFontString ( nil , "OVERLAY" , "GameFontNormal" )
-        UI.MMA_checkbox.Text:SetText( "Nonstop Crafting" )
-        UI.MMA_checkbox.Text:SetPoint( "LEFT" , UI.MMA_checkbox, "RIGHT" , 3 , 0 )
+        UI.MSA_checkbox.Text = UI.MSA_checkbox:CreateFontString ( nil , "OVERLAY" , "GameFontNormal" )
+        UI.MSA_checkbox.Text:SetText( "Nonstop Crafting" )
+        UI.MSA_checkbox.Text:SetPoint( "LEFT" , UI.MSA_checkbox, "RIGHT" , 2 , 0 )
 
         -- Normalize the click area of check button to length of the text
-        UI.MMA_checkbox:SetHitRectInsets ( 0 , 0 - UI.MMA_checkbox.Text:GetWidth() - 2 , 0 , 0 );
+        UI.MSA_checkbox:SetHitRectInsets ( 0 , 0 - UI.MSA_checkbox.Text:GetWidth() - 2 , 0 , 0 );
+
+        -- Ensures this is always to the left of the CreateAllButton, accounting for width of text also
+        UI.MSA_checkbox:SetPoint ( "RIGHT" , ProfessionsFrame.CraftingPage.CreateAllButton , "LEFT" , - math.abs ( select ( 2 , UI.MSA_checkbox:GetHitRectInsets() ) ) - 12 , 0 )
 
         -- Change the setting wether enabled or not
-        UI.MMA_checkbox:SetScript ( "OnClick" , function( self )
-            MCA_save.non_stop = self:GetChecked()
+        UI.MSA_checkbox:SetScript ( "OnClick" , function( self )
+            MSA_save.non_stop = self:GetChecked()
         end)
 
         -- Tooltip
-        UI.MMA_checkbox:SetScript ( "OnEnter" , function( self )
+        UI.MSA_checkbox:SetScript ( "OnEnter" , function( self )
             GameTooltip:SetOwner ( self , "ANCHOR_CURSOR" );
-            GameTooltip:AddLine ( "MCA Limitation: Nonstop will only work for reagents within player bags, not bank tabs." );
+            GameTooltip:AddLine ( "MSA Limitation: Nonstop will only work for reagents within player bags, not bank tabs." );
             GameTooltip:Show();
         end);
 
-        UI.MMA_checkbox:SetScript ( "OnLeave" , function()
+        UI.MSA_checkbox:SetScript ( "OnLeave" , function()
             GameTooltip:Hide()
         end)
 
@@ -81,10 +82,10 @@ end
 -- Purpose:         Only show checkbox when necessary.
 UI.Configure_Visiblity = function( _ , _ , spell_id )
     if spell_id then
-        if UI.Is_Supported_Profession(C_TradeSkillUI.GetBaseProfessionInfo().professionID) and MCA.Crafting.IsMassCraftingSpell(spell_id) then
-            UI.MMA_checkbox:Show()
+        if UI.Is_Supported_Profession(C_TradeSkillUI.GetBaseProfessionInfo().professionID) and MSA.Crafting.IsMassCraftingSpell(spell_id) then
+            UI.MSA_checkbox:Show()
         else
-            UI.MMA_checkbox:Hide()
+            UI.MSA_checkbox:Hide()
         end
     end
 end